Feat: Make Domain layer mapper in member domain #8
